Make the structure clearer
As per feedback on the PR from CS.
