+++ lib/tax_tribunal.rb
@@ -1,17 +1,7 @@
 require 'sinatra'
 require_relative 'tax_tribunal/case'
 require_relative 'tax_tribunal/file'
+require_relative 'tax_tribunal/download'
 
 module TaxTribunal
-  class Downloader < Sinatra::Base
-    configure do
-      set :raise_errors, true
-      set :show_exceptions, false
-    end
-
-    get '/:case_id' do |case_id|
-      @case = Case.new(case_id)
-      erb :show
-    end
-  end
 end
