+++ views/show.erubis
